cgroup: move cgroup setup code to new file
